backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
Pillow==10.0.0
django-colorfield==0.10.1
django-filter==23.2
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
reportlab==4.0.4
flake8==6.1.0
python-dotenv==1.0.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==10.0.0